Const redemption day svn status! ;)

+++ arch/x86_64/include/hack_stdint.h
@@ -31,8 +31,8 @@
   typedef unsigned short uint16_t;
   typedef signed int int32_t;
   typedef unsigned int uint32_t;
-  typedef signed long long int64_t;
-  typedef unsigned long long uint64_t;
+  typedef signed long int64_t;
+  typedef unsigned long uint64_t;
 #endif
 
 #endif
